+++ frontend/src/components/layout/MypageNav.vue
@@ -11,7 +11,7 @@
                                 <router-link to="/user/mypage/orderhistory/">주문내역</router-link>
                             </li>
                             <li class="first">
-                                <router-link to="/user/mypage/mywishlist">관심상품</router-link>
+                                <router-link to="/user/mypage/mywishlist/">관심상품</router-link>
                             </li>
                         </ul>
                     </div>
